BusinessLicenseNavigator/pyproject.toml
[project]
name = "businesslicensenavigator"
version = "0.1.0"
description = "AI-powered business license guidance system with Delaware RAG integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework and UI
    "streamlit>=1.47.1",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    
    # HTTP and Web Scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    
    # AI and Language Models
    "google-generativeai>=0.3.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langchain-community>=0.0.20",
    "langchain-groq>=0.3.6",
    
    # MCP (Model Context Protocol) Components
    "mcp>=1.0.0",
    "mcp-use>=1.3.7",
    
    # Vector Database and Embeddings
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.7.0",
    "numpy>=1.24.0",
    
    # Configuration and Environment
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    
    # Utilities
    "typing-extensions>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
